pyproject.toml
[project]
name = "sample-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "google-adk>=1.0.0",
    "isort>=6.0.1",
    "requests>=2.32.3",
    "rich",
    "mcp[cli]"
]
